refactor(generator/test): split "used-by" tests
